photo_downloader/requirements.txt
instabot==0.4.5
Pillow==6.2.0
requests==2.22.0
